+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.1] - 29/09/2020.
+
+* Add comments;
+* Change utils to classes and refactor package folders;
+
+* TODO: calculate the Secondary Color based on the Primary Color when the Secondary is not informed.
+
 ## [0.1.0] - 29/09/2020.
 
 * Init Project;
@@ -5,4 +12,4 @@
 * Add README.md;
 * Add tests;
 
-* TODO; calculate the Secondary Color based on the Primary Color when the Secondary is not informed.
+* TODO: calculate the Secondary Color based on the Primary Color when the Secondary is not informed.